Handling and protecting multiple cases

+++ bonus/server_bonus.c
@@ -6,28 +6,73 @@
 /*   By: noaziki <noaziki@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/18 16:44:14 by noaziki           #+#    #+#             */
-/*   Updated: 2025/04/18 17:59:50 by noaziki          ###   ########.fr       */
+/*   Updated: 2025/04/20 20:48:33 by noaziki          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minitalk_bonus.h"
 
+static int	check_bytes(char *arr)
+{
+	if ((arr[0] & 128) == 0)
+		return (1);
+	if ((arr[0] & 224) == 192)
+		return (2);
+	if ((arr[0] & 240) == 224)
+		return (3);
+	if ((arr[0] & 248) == 240)
+		return (4);
+	return (1);
+}
+
+static void	handle_byte(char *arr, int *bit, int *byte)
+{
+	static int	exp;
+
+	(*byte)++;
+	if (*byte == 1)
+		exp = check_bytes(arr);
+	if (*byte == 1 && exp == 1)
+	{
+		write(1, &arr[0], 1);
+		*byte = 0;
+		ft_bzero(arr, 1);
+	}
+	if (*byte == exp && exp != 1)
+	{
+		write(1, arr, exp);
+		*byte = 0;
+		ft_bzero(arr, 4);
+	}
+	*bit = 0;
+}
+
 void	handle_sig(int sig, siginfo_t *info, void *context)
 {
-	static int	bit = 0;
-	static char	c = 0;
+	static char		arr[4];
+	static int		bit;
+	static int		byte;
+	static pid_t	id;
 
 	(void)context;
+	if (id == 0)
+		id = info->si_pid;
+	if (id != info->si_pid)
+	{
+		id = 0;
+		bit = 0;
+		byte = 0;
+		ft_bzero(arr, 4);
+	}
 	if (sig == SIGUSR1)
-		c |= (1 << bit);
+		arr[byte] |= (1 << bit);
 	bit++;
 	if (bit == 8)
 	{
-		write(1, &c, 1);
-		bit = 0;
-		c = 0;
+		if (arr[0] == '\0')
+			kill(info->si_pid, SIGUSR1);
+		handle_byte(arr, &bit, &byte);
 	}
-	kill(info->si_pid, SIGUSR1);
 }
 
 int	main(void)
@@ -41,7 +86,6 @@ int	main(void)
 	ft_putstr_fd("\n", 1);
 	sa.sa_sigaction = &handle_sig;
 	sa.sa_flags = SA_SIGINFO;
-	sigemptyset(&sa.sa_mask);
 	sigaction(SIGUSR1, &sa, NULL);
 	sigaction(SIGUSR2, &sa, NULL);
 	while (1)
